Image_Recoginition with pre-trained model
In this code, we first load the pre-trained VGG16 model using the vgg16.VGG16() function. We then load an image, resize it to the input size of the model, and convert it to a numpy array using the image.load_img() and image.img_to_array() functions. We preprocess the image by normalizing the pixel values and converting the color format from RGB to BGR using the vgg16.preprocess_input() function.

We then use the model to predict the label of the image using the model.predict() function. The predictions variable contains the probability distribution over the 1000 classes of the ImageNet dataset. We decode the predictions to obtain the human-readable label using the vgg16.decode_predictions() function. The top parameter specifies how many top predictions to return, and we choose top=1 to obtain the most likely label. Finally, we print the label using the print() function.