Video length functionality in video screen
This script adds a Dropdown component to the video camera screen and populates it with three options for the video length: 15 seconds, 30 seconds, and 60 seconds. The initial video length is set to 15 seconds. In the Update() function, the script checks which option is currently selected in the dropdown menu and sets the video length variable accordingly. You can use this variable to control the length of the video recording in your video camera script.

Taiga Link: https://tree.taiga.io/project/jmpatel3998-ser517-ar-map-builder/task/54