Created the ButtonBackend.cs
This script allows the user to measure the length of the camera by clicking and dragging on the screen. When the user clicks the button, the OnClick method is called. If the user is not currently measuring the camera length, the button text is changed to "Finish Measuring" and the lengthText is set to "Measuring camera length...". If the user is already measuring, the button text is changed back to "Measure Camera Length" and the camera length is calculated.

The Update method is called every frame and checks if the user is currently measuring the camera length. If so, it checks for mouse button clicks and updates the start and end positions accordingly.

Note that this script assumes that the camera is facing in the positive z direction. If your camera is facing a different direction, you may need to modify the script accordingly. Additionally, the script assumes that there are no obstacles between the camera and the start and end positions. If there are obstacles, the raycast may hit the obstacle instead of the position you intended, resulting in incorrect measurements.

Taiga Link: https://tree.taiga.io/project/jmpatel3998-ser517-ar-map-builder/task/59