Created ObjectTargetEdit.cs
This code creates a script that implements the ITrackableEventHandler interface, which allows you to handle object tracking events such as when the object is detected or lost. In the Start function, the script registers the object target's trackable event handler. The OnTrackableStateChanged function is called whenever the object's tracking state changes, and depending on the state, it calls either OnTrackingFound or OnTrackingLost functions.

Taiga link: https://tree.taiga.io/project/jmpatel3998-ser517-ar-map-builder/task/77